+++ crates/chain-config/src/config/state/reader.rs
@@ -1,11 +1,21 @@
 use std::fmt::Debug;
 
-use fuel_core_storage::{structured_storage::TableWithBlueprint, Mappable};
-use fuel_core_types::{blockchain::primitives::DaBlockHeight, fuel_types::BlockHeight};
+use fuel_core_storage::{
+    structured_storage::TableWithBlueprint,
+    Mappable,
+};
+use fuel_core_types::{
+    blockchain::primitives::DaBlockHeight,
+    fuel_types::BlockHeight,
+};
 use itertools::Itertools;
 
 use crate::{
-    config::table_entry::TableEntry, AsTable, ChainConfig, StateConfig, MAX_GROUP_SIZE,
+    config::table_entry::TableEntry,
+    AsTable,
+    ChainConfig,
+    StateConfig,
+    MAX_GROUP_SIZE,
 };
 
 pub struct Groups<T: Mappable> {