+++ crates/fuel-core/src/service/genesis/importer.rs
@@ -1,35 +1,61 @@
 use self::{
     import_task::ImportTable,
-    progress::{MultipleProgressReporter, ProgressReporter, Target},
+    progress::{
+        MultipleProgressReporter,
+        ProgressReporter,
+        Target,
+    },
 };
 
 use super::task_manager::TaskManager;
 mod import_task;
 mod on_chain;
 mod progress;
-use std::{io::IsTerminal, marker::PhantomData};
+use std::{
+    io::IsTerminal,
+    marker::PhantomData,
+};
 
 use crate::{
     combined_database::CombinedDatabase,
-    database::database_description::{off_chain::OffChain, on_chain::OnChain},
+    database::database_description::{
+        off_chain::OffChain,
+        on_chain::OnChain,
+    },
     graphql_api::storage::{
         coins::OwnedCoins,
         contracts::ContractsInfo,
         messages::OwnedMessageIds,
-        transactions::{OwnedTransactions, TransactionStatuses},
+        transactions::{
+            OwnedTransactions,
+            TransactionStatuses,
+        },
     },
 };
-use fuel_core_chain_config::{AsTable, SnapshotReader, StateConfig, TableEntry};
+use fuel_core_chain_config::{
+    AsTable,
+    SnapshotReader,
+    StateConfig,
+    TableEntry,
+};
 use fuel_core_services::StateWatcher;
 use fuel_core_storage::{
     kv_store::StorageColumn,
     structured_storage::TableWithBlueprint,
     tables::{
-        Coins, ContractsAssets, ContractsLatestUtxo, ContractsRawCode, ContractsState,
-        Messages, Transactions,
+        Coins,
+        ContractsAssets,
+        ContractsLatestUtxo,
+        ContractsRawCode,
+        ContractsState,
+        Messages,
+        Transactions,
     },
 };
-use fuel_core_types::{blockchain::primitives::DaBlockHeight, fuel_types::BlockHeight};
+use fuel_core_types::{
+    blockchain::primitives::DaBlockHeight,
+    fuel_types::BlockHeight,
+};
 
 use tracing::Level;