+++ crates/fuel-core/src/service/genesis/importer/import_task.rs
@@ -1,17 +1,30 @@
-use fuel_core_chain_config::{Groups, TableEntry};
+use fuel_core_chain_config::{
+    Groups,
+    TableEntry,
+};
 use fuel_core_storage::{
     kv_store::StorageColumn,
     structured_storage::TableWithBlueprint,
-    transactional::{StorageTransaction, WriteTransaction},
-    Mappable, StorageAsRef, StorageInspect,
+    transactional::{
+        StorageTransaction,
+        WriteTransaction,
+    },
+    Mappable,
+    StorageAsRef,
+    StorageInspect,
 };
 
 use crate::{
     database::{
         database_description::{
-            off_chain::OffChain, on_chain::OnChain, DatabaseDescription,
+            off_chain::OffChain,
+            on_chain::OnChain,
+            DatabaseDescription,
         },
-        genesis_progress::{GenesisMetadata, GenesisProgressMutate},
+        genesis_progress::{
+            GenesisMetadata,
+            GenesisProgressMutate,
+        },
         Database,
     },
     service::genesis::task_manager::CancellationToken,
@@ -103,7 +116,7 @@ where
 
             off_chain_tx.commit()?;
 
-            reporter.set_progress(index as u64);
+            reporter.set_progress(u64::try_from(index).unwrap_or(u64::MAX));
 
             Ok(())
         });
@@ -121,40 +134,78 @@ mod tests {
             genesis_progress::GenesisProgressInspect,
         },
         service::genesis::{
-            importer::{import_task::import_entries, progress::ProgressReporter},
+            importer::{
+                import_task::import_entries,
+                progress::ProgressReporter,
+            },
             task_manager::CancellationToken,
         },
     };
     use std::{
-        sync::{Arc, Mutex},
+        sync::{
+            Arc,
+            Mutex,
+        },
         time::Duration,
     };
 
-    use anyhow::{anyhow, bail};
-    use fuel_core_chain_config::{Groups, Randomize, TableEntry};
+    use anyhow::{
+        anyhow,
+        bail,
+    };
+    use fuel_core_chain_config::{
+        Groups,
+        Randomize,
+        TableEntry,
+    };
     use fuel_core_storage::{
         column::Column,
-        iter::{BoxedIter, IterDirection, IterableStore},
-        kv_store::{KVItem, KeyValueInspect, StorageColumn, Value},
+        iter::{
+            BoxedIter,
+            IterDirection,
+            IterableStore,
+        },
+        kv_store::{
+            KVItem,
+            KeyValueInspect,
+            StorageColumn,
+            Value,
+        },
         structured_storage::TableWithBlueprint,
         tables::Coins,
-        transactional::{Changes, StorageTransaction},
-        Result as StorageResult, StorageAsMut, StorageAsRef, StorageInspect,
+        transactional::{
+            Changes,
+            StorageTransaction,
+        },
+        Result as StorageResult,
+        StorageAsMut,
+        StorageAsRef,
+        StorageInspect,
     };
     use fuel_core_types::{
-        entities::coins::coin::{CompressedCoin, CompressedCoinV1},
+        entities::coins::coin::{
+            CompressedCoin,
+            CompressedCoinV1,
+        },
         fuel_tx::UtxoId,
         fuel_types::BlockHeight,
     };
-    use rand::{rngs::StdRng, SeedableRng};
+    use rand::{
+        rngs::StdRng,
+        SeedableRng,
+    };
 
     use crate::{
         combined_database::CombinedDatabase,
         database::{
             database_description::on_chain::OnChain,
-            genesis_progress::GenesisProgressMutate, Database,
+            genesis_progress::GenesisProgressMutate,
+            Database,
         },
-        state::{in_memory::memory_store::MemoryStore, TransactableStorage},
+        state::{
+            in_memory::memory_store::MemoryStore,
+            TransactableStorage,
+        },
     };
 
     use super::ImportTable;