+++ crates/fuel-core/src/service/genesis/importer/on_chain.rs
@@ -1,17 +1,26 @@
 use std::borrow::Cow;
 
-use super::{import_task::ImportTable, Handler};
+use super::{
+    import_task::ImportTable,
+    Handler,
+};
 use crate::{
     database::{
         balances::BalancesInitializer,
-        database_description::{off_chain::OffChain, on_chain::OnChain},
+        database_description::{
+            off_chain::OffChain,
+            on_chain::OnChain,
+        },
         state::StateInitializer,
         Database,
     },
     graphql_api::{
         storage::{
             blocks::FuelBlockIdsToHeights,
-            transactions::{OwnedTransactions, TransactionStatuses},
+            transactions::{
+                OwnedTransactions,
+                TransactionStatuses,
+            },
         },
         worker_service,
     },
@@ -20,8 +29,13 @@ use anyhow::anyhow;
 use fuel_core_chain_config::TableEntry;
 use fuel_core_storage::{
     tables::{
-        Coins, ContractsAssets, ContractsLatestUtxo, ContractsRawCode, ContractsState,
-        Messages, Transactions,
+        Coins,
+        ContractsAssets,
+        ContractsLatestUtxo,
+        ContractsRawCode,
+        ContractsState,
+        Messages,
+        Transactions,
     },
     transactional::StorageTransaction,
     StorageAsMut,
@@ -29,7 +43,10 @@ use fuel_core_storage::{
 use fuel_core_types::{
     self,
     blockchain::primitives::DaBlockHeight,
-    entities::{coins::coin::Coin, Message},
+    entities::{
+        coins::coin::Coin,
+        Message,
+    },
     fuel_types::BlockHeight,
     services::executor::Event,
 };